+++ src/extension.ts
@@ -1,37 +1,31 @@
 import { spawn } from 'child_process';
-import { commands, ExtensionContext, OutputChannel, Uri, window, workspace, WorkspaceConfiguration } from 'vscode';
-import { buildCommandFromConfig, getActiveDocumentPath, getWorkspaceRootPath } from './util';
+import { commands, ExtensionContext, Task, tasks, TaskScope, Uri, window, workspace } from 'vscode';
+import { getFormatTasks, PintTaskProvider } from './pintTaskProvider';
+import { buildCommandFromConfig, getActiveDocumentPath, getOutputChannel, getWorkspaceConfig, getWorkspaceRootPath } from './util';
 const pkg = require('../package.json');
 
-let outputChannel: OutputChannel | null;
-
-async function format(file: Uri, config: WorkspaceConfiguration) {
+async function format(file: Uri) {
   const filePath = workspace.asRelativePath(file.path);
   let command: string | undefined;
-  const commandParts = await buildCommandFromConfig(filePath, config);
+  const commandParts = await buildCommandFromConfig(filePath);
 
   if (commandParts === false || !(command = commandParts.shift())) {
-    outputChannel?.appendLine(`Something went wrong! Executable does not exists or lacks permissions. Please check before create an issue on ${pkg.bugs.url}`);
+    getOutputChannel().appendLine(`Something went wrong! Executable does not exists or lacks permissions. Please check before create an issue on ${pkg.bugs.url}`);
 
     return;
   }
   
-  // Use this for debugging purposes...
-  outputChannel?.appendLine(`Formatting file "${filePath}" using command "${command} ${commandParts.join(' ')}"`);
+  getOutputChannel().appendLine(
+    (filePath ? `Formatting file "${filePath}"` : 'Formatting workspace folder') + ` using command "${command} ${commandParts.join(' ')}"`
+  );
 
+  // TODO: Output stdout, etc...?
   const exec = spawn(command, commandParts, {
     cwd: getWorkspaceRootPath()
   });
 }
 
-function getWorkspaceConfig() {
-  return workspace.getConfiguration('laravel-pint');
-}
-
 export function activate(context: ExtensionContext) {
-  outputChannel = window.createOutputChannel('Laravel Pint');
-  let config = getWorkspaceConfig();
-
   context.subscriptions.push(commands.registerCommand('laravel-pint.format', () => {
     if (window.activeTextEditor?.document.languageId !== 'php') {
       return;
@@ -40,16 +34,20 @@ export function activate(context: ExtensionContext) {
     const activeEditorFile = getActiveDocumentPath();
 
     if (activeEditorFile) {
-      format(activeEditorFile, config);
+      format(activeEditorFile);
     }
   }));
 
-  context.subscriptions.push(workspace.onDidChangeConfiguration(() => {
-    config = getWorkspaceConfig();
+  context.subscriptions.push(commands.registerCommand('laravel-pint.formatProject', async () => {
+    const extensionTasks = await getFormatTasks();
+    
+    tasks.executeTask(extensionTasks[0]);
   }));
 
+  context.subscriptions.push(tasks.registerTaskProvider(PintTaskProvider.PintType, new PintTaskProvider));
+
   context.subscriptions.push(workspace.onWillSaveTextDocument((e) => {
-    const saveEnabled = config.get<boolean>('formatOnSave');
+    const saveEnabled = getWorkspaceConfig<boolean>('formatOnSave');
     
     if (!saveEnabled || e.document.languageId !== 'php') {
       return;
@@ -58,17 +56,11 @@ export function activate(context: ExtensionContext) {
     const activeEditorFile = getActiveDocumentPath();
 
     if (activeEditorFile) {
-      format(activeEditorFile, config);
+      format(activeEditorFile);
     }
   }));
 }
 
-// this method is called when your extension is deactivated
 export function deactivate() {
-  if (outputChannel) {
-    outputChannel.clear();
-    outputChannel.dispose();
-  }
-
-  outputChannel = null;
+  getOutputChannel(true);
 }