+++ src/util/index.ts
@@ -1,2 +1,3 @@
 export * from "./command";
 export * from "./filesystem";
+export * from "./misc";