+++ CHANGELOG.md
@@ -6,11 +6,18 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+## [0.7.0] - 2022-06-30
+
+### Added
+
+- Format workspace files command.
+- Format task provider.
+
 ## [0.6.1] - 2022-06-29
 
 ### Fixed
 
-- Regression on alerts when projects aren't using Laravel Pint (#1)
+- Regression on alerts when projects aren't using Laravel Pint (#1).
 
 ## [0.6.0] - 2022-06-28